+++ src/__init__.py
@@ -23,11 +23,11 @@ in python3
 """
 
 __author__ = "Dr. Ralf Antonius Timmermann"
-__copyright__ = ("Copyright (c) 2023-2024, Dr. Ralf Antonius Timmermann "
+__copyright__ = ("Copyright (c) 2023-2025, Dr. Ralf Antonius Timmermann "
                  "All rights reserved.")
 __credits__ = ""
 __license__ = "BSD-3-Clause"
-__version__ = "0.2.1"
+__version__ = "1.0.0"
 __maintainer__ = "Dr. Ralf Antonius Timmermann"
 __email__ = "ralf.timmermann@gmx.de"
 __status__ = "Prod"